# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Book API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/circles": {
            "get": {
                "description": "Every Circle the authenticated user is currently an active member of",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "List my circles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCirclesResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new private Circle; the caller becomes its admin",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "Create a circle",
                "parameters": [
                    {
                        "description": "Create circle request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CreateCircleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/circles/{id}": {
            "get": {
                "description": "Get a single Circle; requires the caller to be an active member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "Get a circle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            },
            "put": {
                "description": "Full-representation update of a Circle; admin-only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "Update a circle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update circle request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft-deletes a Circle; admin-only",
                "tags": [
                    "circles"
                ],
                "summary": "Dissolve a circle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
        "/circles/{id}/members": {
            "get": {
                "description": "The Circle's active roster; requires the caller to be an active member",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "List a circle's members",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCircleMembersResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
        "/circles/{id}/members/{userId}": {
            "delete": {
                "description": "Self-leave when userId is the caller's own id, otherwise an admin-forced removal",
                "tags": [
                    "circles"
                ],
                "summary": "Remove a circle member",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            },
            "patch": {
                "description": "Admin-only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "Change a circle member's permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update permissions request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleMemberPermissionsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleMemberResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/circles/{id}/members/{userId}/role": {
            "patch": {
                "description": "Admin-only",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "circles"
                ],
                "summary": "Change a circle member's role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Circle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target user ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update role request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleMemberRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleMemberResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse": {
            "type": "object",
            "properties": {
                "can_capture": {
                    "type": "boolean"
                },
                "can_invite": {
                    "type": "boolean"
                },
                "circle_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "joined_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "user_id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleResponse": {
            "type": "object",
            "properties": {
                "color_hex": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "created_by_user_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dissolved_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                },
                "image_path": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "NgodingVareng"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CreateCircleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "color_hex": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "image_path": {
                    "type": "string",
                    "maxLength": 1024
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "NgodingVareng"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCircleMembersResponse": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse"
                    }
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCirclesResponse": {
            "type": "object",
            "properties": {
                "circles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleResponse"
                    }
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleMemberPermissionsRequest": {
            "type": "object",
            "properties": {
                "can_capture": {
                    "type": "boolean"
                },
                "can_invite": {
                    "type": "boolean"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleMemberRoleRequest": {
            "type": "object",
            "required": [
                "role"
            ],
            "properties": {
                "role": {
                    "type": "string",
                    "enum": [
                        "admin",
                        "member"
                    ],
                    "example": "admin"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.UpdateCircleRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "color_hex": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "image_path": {
                    "type": "string",
                    "maxLength": 1024
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1,
                    "example": "NgodingVareng"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-any": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleMemberResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleMemberResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_CircleResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.CircleResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCircleMembersResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCircleMembersResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "github_com_ngodingvareng_memoria_internal_delivery_rest_dto.WebResponse-github_com_ngodingvareng_memoria_internal_delivery_rest_dto_ListCirclesResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/github_com_ngodingvareng_memoria_internal_delivery_rest_dto.ListCirclesResponse"
                },
                "errors": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
    }
}